pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Leandro Mirante <leandro.mirante@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.1"}
Pillow = "^9.4.0"
django-cors-headers = "^3.13.0"
algoliasearch-django = ">=2.0,<3.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
six = "^1.16.0"
drf-yasg = "^1.21.4"
flake8 = "^6.0.0"
black = "^22.12.0"
django-jazzmin = "^2.6.0"
celery = "^5.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
